requirements.txt
autopep8==1.5.3
bcrypt==3.1.7
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
dataclasses==0.7
fastapi==0.58.1
h11==0.9.0
httptools==0.1.1
idna==2.10
passlib==1.7.2
pkg-resources==0.0.0
pycodestyle==2.6.0
pycparser==2.20
pydantic==1.5.1
PyJWT==1.7.1
python-dotenv==0.13.0
python-multipart==0.0.5
requests==2.24.0
six==1.15.0
starlette==0.13.4
toml==0.10.1
urllib3==1.25.9
uvicorn==0.11.5
uvloop==0.14.0
websockets==8.1
